Cleanup
* move functions from `helper` that were only used in `inhibitory_neuron_densities_optimization`
* removed unused `with_descendants` argument from `_compute_region_cell_counts`
* made `compute_region_volumes` faster
